Completed emil notification functionality
Separated Email functionality by adding a new project that runs a console
application. This should be set up to run once a day with windows task
scheduler. This seemed to be the best way to implement email notifications
since now the email portion of the app won't be able to throw an exception
and crash the app.